Moved ClusterClientHandler static classes to individual files to comply with CodeFactor.
